requirements.txt
watchdog
requests
python-box
pyyaml
